test: keep fallow focused on tvOS changes
